pyproject.toml
[project]
name = "langgraph-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "langgraph-cli[inmem]>=0.4.12",
    "langsmith>=0.7.3",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
]
